+++ changelog.md
@@ -3,7 +3,7 @@
 Format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 Adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ Unreleased ]
+## [ Unreleased(v0.2.3) ]
 
 ### Removed
 - Removed website, moved to single README.md file.